requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
celery
redis
supabase
httpx
python-multipart
